chore: moved IPC queue to its own module #7
