+++ CM7/Core/Inc/steering/screen.h
@@ -1,0 +1,7 @@
+#ifndef SCREEN_H
+#define SCREEN_H
+
+#define SCREEN_WIDTH (800U)
+#define SCREEN_HEIGHT (480U)
+
+#endif // SCREEN_H
