+++ CM7/Core/Src/main.c
@@ -35,7 +35,8 @@
 #include "fsm.h"
 #include "eagletrt.h"
 #include "input-events-api.h"
-#include "inputs-shared-api.h"
+#include "ipc-api.h"
+#include "shared.h"
 
 /* USER CODE END Includes */
 
@@ -74,7 +75,7 @@ static void MPU_Config(void);
 
 void HAL_HSEM_FreeCallback(uint32_t SemMask) {
     if (SemMask & (1 << HSEM_INPUT_ID)) {
-        inputs_shared_api_read_and_process_all(&ipc_input, input_events_api_handle_event);
+        ipc_api_read_and_process_all(&ipc_input, input_events_api_handle_event);
         HAL_HSEM_ActivateNotification(__HAL_HSEM_SEMID_TO_MASK(HSEM_INPUT_ID));
     }
 }
