+++ Common/Inc/shared.h
@@ -1,0 +1,9 @@
+#ifndef SHARED_H
+#define SHARED_H
+
+/*!
+ * \brief HSEM ID used by CM4 to notify CM7 about input events.
+ */
+#define HSEM_INPUT_ID (1U)
+
+#endif // SHARED_H
