+++ Common/Inc/steering/inputs-shared.h
@@ -15,16 +15,6 @@
 #include <stdint.h>
 
 /*!
- * \brief HSEM ID used by CM4 to notify CM7 about input events.
- */
-#define HSEM_INPUT_ID (1U)
-
-/*!
- * \brief Size of the IPC input event ring buffer.
- */
-#define IPC_INPUT_QUEUE_SIZE 64
-
-/*!
  * \brief Enumeration of input event types
  */
 enum InputEventType {
@@ -83,15 +73,6 @@ struct InputEvent {
             enum ButtonID button_id; /*!< Identifier for the button */
         } button;
     };
-};
-
-/*!
- * \brief Lock-free single-producer single-consumer ring buffer for input events.
- */
-struct IPCInputQueue {
-    uint32_t write_idx;                             /*!< Next slot to write (owned by CM4) */
-    uint32_t read_idx;                              /*!< Next slot to read  (owned by CM7) */
-    struct InputEvent events[IPC_INPUT_QUEUE_SIZE]; /*!< Ring buffer storage */
 };
 
 #endif // INPUTS_SHARED_H
