+++ Common/Src/steering/ipc-api.c
@@ -1,6 +1,6 @@
-#include "inputs-shared-api.h"
+#include "ipc-api.h"
 
-void inputs_shared_api_read_and_process_all(struct IPCInputQueue *queue, void (*callback)(struct InputEvent *ev)) {
+void ipc_api_read_and_process_all(struct IPCInputQueue *queue, void (*callback)(struct InputEvent *ev)) {
     while (queue->read_idx != queue->write_idx) {
         struct InputEvent ev =
             queue->events[queue->read_idx];
@@ -12,7 +12,7 @@ void inputs_shared_api_read_and_process_all(struct IPCInputQueue *queue, void (*
     }
 }
 
-bool inputs_shared_api_push_event(struct IPCInputQueue *queue, struct InputEvent *ev, void (*dmb_callback)(void)) {
+bool ipc_api_push_event(struct IPCInputQueue *queue, struct InputEvent *ev, void (*dmb_callback)(void)) {
     uint32_t next = (queue->write_idx + 1) % IPC_INPUT_QUEUE_SIZE;
 
     if (next == queue->read_idx) {
