+++ platformio.ini
@@ -27,6 +27,7 @@ build_flags =
     -ICM7/FATFS/App
     -IMiddlewares/Third_Party/FatFs/src
     -ICommon/Inc
+    -ICommon/Inc/steering
 build_src_filter =
     +<CM7/Core/Src/>
     +<Drivers/STM32H7xx_HAL_Driver/Src>
@@ -75,6 +76,7 @@ build_flags =
     -IDrivers/CMSIS/Device/ST/STM32H7xx/Include
     -IDrivers/CMSIS/Include
     -ICommon/Inc
+    -ICommon/Inc/steering
 build_src_filter =
     +<CM4/Core/Src/>
     +<Drivers/STM32H7xx_HAL_Driver/Src>
